nautilus-directory: handle change events
for files that were still not known to nautilus because they were in
the midst of being added to Nautilus. This happens when eg. inotify
fires two consecutive NEW + CHANGE events for a new file, and it's
important to handle the CHANGE event as the file may have changed
its properties with respect the ones reported at the time the NEW
event fired.

A case this happens is for mountpoint directories of
removable devices, as seen in the referenced bug below.

We now queue CHANGE events received for files unknown to Nautilus only
when the parent folder is currently adding new files. When the folder
finish adding the new files, we then process the list of queued files.
If there are still files unknown to Nautilus in this list, they will
be ignored as before.

https://bugzilla.gnome.org/show_bug.cgi?id=703179
Issue #1576