file-operations: Do not offer skipping when extracting one file
In case of extraction failure, the "Skip" and "Cancel" actions are offered
everytime, but skipping doesn't make sense when extracting one file only.
Let's use the same approach as it is used also for other operations, which
is based on total number of files and remaining files. Also the "Skip All"
action will be offered as a side-effect of this change.